Implement review feedback. + Moved the collection of analyzer results to the mount function of the AnalyzerResults component. + Adjusted to the dark & light theme + Bug fixes & clean-up
